tons of plugin refactoring
moved plugins to new lib/plugins folder, moved most plugins logic out of bundle source, initial steps to be able to toggle visualisations on the client for enabled plugins
