+++ bower.json
@@ -7,8 +7,9 @@
     "d3": "3.4.3",
     "jquery": "2.1.0",
     "jQuery-Storage-API": "~1.7.2",
-    "tipsy-jmalonzo": "~1.0.1",
-    "jsSHA": "~1.5.0"
+    "jsSHA": "~1.5.0",
+    "lodash": "~3.9.0",
+    "tipsy-jmalonzo": "~1.0.1"
   },
   "resolutions": {
     "jquery": "2.1.0"
