+++ bundle/index.js
@@ -5,7 +5,7 @@ var browserify_express = require('browserify-express');
 function bundle() {
   return browserify_express({
     entry: __dirname + '/bundle.source.js',
-    watch: __dirname + '/../lib/',
+    watch: __dirname + '/../lib/plugins/',
     mount: '/public/js/bundle.js',
     verbose: true,
     //minify: true,
