+++ lib/pebble.js
@@ -13,7 +13,7 @@ var DIRECTIONS = {
 , 'RATE OUT OF RANGE': 9
 };
 
-var iob = require("./iob")();
+var iob = require("./plugins/iob")();
 var async = require('async');
 var units = require('./units')();
