+++ lib/plugins/iob.js
@@ -1,8 +1,7 @@
 'use strict';
 
-function getData()
-{
-	return calcTotal(this.env.treatments,this.env.profile,this.env.time);
+function getData() {
+  return calcTotal(this.env.treatments,this.env.profile,this.env.time);
 }
 
 function calcTotal(treatments, profile, time) {
@@ -50,13 +49,13 @@ function calcTreatment(treatment, profile, time) {
     var minAgo = scaleFactor * (time - bolusTime) / 1000 / 60;
 
     if (minAgo < peak) {
-      var x = minAgo / 5 + 1;
-      iobContrib = treatment.insulin * (1 - 0.001852 * x * x + 0.001852 * x);
+      var x1 = minAgo / 5 + 1;
+      iobContrib = treatment.insulin * (1 - 0.001852 * x1 * x1 + 0.001852 * x1);
       activityContrib = sens * treatment.insulin * (2 / dia / 60 / peak) * minAgo;
 
     } else if (minAgo < 180) {
-      var x = (minAgo - 75) / 5;
-      iobContrib = treatment.insulin * (0.001323 * x * x - .054233 * x + .55556);
+      var x2 = (minAgo - 75) / 5;
+      iobContrib = treatment.insulin * (0.001323 * x2 * x2 - .054233 * x2 + .55556);
       activityContrib = sens * treatment.insulin * (2 / dia / 60 - (minAgo - peak) * 2 / dia / 60 / (60 * dia - peak));
     } else {
       iobContrib = 0;
@@ -73,24 +72,23 @@ function calcTreatment(treatment, profile, time) {
 }
 
 function updateVisualisation() {
-	
-	var pill = this.currentDetails.find('span.pill.iob');
+  var pill = this.currentDetails.find('span.pill.iob');
 
-	if (!pill || pill.length == 0) {
-		pill = $('<span class="pill iob"><label>IOB</label><em></em></span>');
-		this.currentDetails.append(pill);
-	}
-		
-	pill.find('em').text(this.iob.display + 'U');
-	
+  if (!pill || pill.length == 0) {
+    pill = $('<span class="pill iob"><label>IOB</label><em></em></span>');
+    this.currentDetails.append(pill);
+  }
+
+  pill.find('em').text(this.iob.display + 'U');
 }
 
 
 function IOB(pluginBase) {
 
   if (pluginBase) { pluginBase.call(this); }
-  
+
   return {
+    label: 'Insulin-on-Board',
     calcTotal: calcTotal,
     getData: getData,
     updateVisualisation: updateVisualisation,
