+++ static/index.html
@@ -132,6 +132,9 @@
                         <dd><input type="radio" name="theme-browser" id="theme-default-browser" value="default" checked /><label for="theme-default-browser">Default</label></dd>
                         <dd><input type="radio" name="theme-browser" id="theme-colors-browser" value="colors" /><label for="theme-colors-browser">Colors</label></dd>
                     </dl>
+                    <dl id="enable-plugins" class="toggle">
+                        <dt>Enable Plugins</dt>
+                    </dl>
                 </fieldset>
 
                 <div class="actions">
